Move obsolete move generator methods to test project
