+++ tests/Lynx.Test/MoveGeneration/ReferenceMoveGenerator.cs
@@ -1,0 +1,219 @@
+﻿using Lynx.Model;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Lynx.Test.MoveGeneration;
+public static class ReferenceMoveGenerator
+{
+    /// <summary>
+    /// Indexed by <see cref="Piece"/>.
+    /// Checks are not considered
+    /// </summary>
+    private static readonly Func<int, BitBoard, BitBoard>[] _pieceAttacks =
+    [
+        (int origin, BitBoard _) => Attacks.PawnAttacks[(int)Side.White][origin],
+        (int origin, BitBoard _) => Attacks.KnightAttacks[origin],
+        Attacks.BishopAttacks,
+        Attacks.RookAttacks,
+        Attacks.QueenAttacks,
+        (int origin, BitBoard _) => Attacks.KingAttacks[origin],
+
+        (int origin, BitBoard _) => Attacks.PawnAttacks[(int)Side.Black][origin],
+        (int origin, BitBoard _) => Attacks.KnightAttacks[origin],
+        Attacks.BishopAttacks,
+        Attacks.RookAttacks,
+        Attacks.QueenAttacks,
+        (int origin, BitBoard _) => Attacks.KingAttacks[origin],
+    ];
+
+    /// <summary>
+    /// Sames as <see cref="GeneratePawnMoves(Position, int, bool)"/> but returning them
+    /// </summary>
+    /// <param name="position"></param>
+    /// <param name="offset"></param>
+    /// <param name="capturesOnly"></param>
+    /// <returns></returns>
+    [Obsolete("Mostly for reference")]
+    internal static IEnumerable<Move> GeneratePawnMovesForReference(Position position, int offset, bool capturesOnly = false)
+    {
+        int sourceSquare, targetSquare;
+
+        var piece = (int)Piece.P + offset;
+        var pawnPush = +8 - ((int)position.Side * 16);          // position.Side == Side.White ? -8 : +8
+        int oppositeSide = Utils.OppositeSide(position.Side);   // position.Side == Side.White ? (int)Side.Black : (int)Side.White
+        var bitboard = position.PieceBitBoards[piece];
+
+        while (bitboard != default)
+        {
+            sourceSquare = bitboard.GetLS1BIndex();
+            bitboard.ResetLS1B();
+
+            var sourceRank = (sourceSquare >> 3) + 1;
+
+            if (sourceRank == 1 || sourceRank == 8)
+            {
+                throw new($"There's a non-promoted {position.Side} pawn in rank {sourceRank}");
+            }
+
+            // Pawn pushes
+            var singlePushSquare = sourceSquare + pawnPush;
+            if (!position.OccupancyBitBoards[2].GetBit(singlePushSquare))
+            {
+                // Single pawn push
+                var targetRank = (singlePushSquare >> 3) + 1;
+                if (targetRank == 1 || targetRank == 8)  // Promotion
+                {
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, singlePushSquare, piece, promotedPiece: (int)Piece.Q + offset);
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, singlePushSquare, piece, promotedPiece: (int)Piece.R + offset);
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, singlePushSquare, piece, promotedPiece: (int)Piece.N + offset);
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, singlePushSquare, piece, promotedPiece: (int)Piece.B + offset);
+                }
+                else if (!capturesOnly)
+                {
+                    yield return MoveExtensions.Encode(sourceSquare, singlePushSquare, piece);
+                }
+
+                // Double pawn push
+                // Inside of the if because singlePush square cannot be occupied either
+                if (!capturesOnly)
+                {
+                    var doublePushSquare = sourceSquare + (2 * pawnPush);
+                    if (!position.OccupancyBitBoards[2].GetBit(doublePushSquare)
+                        && ((sourceRank == 2 && position.Side == Side.Black) || (sourceRank == 7 && position.Side == Side.White)))
+                    {
+                        yield return MoveExtensions.EncodeDoublePawnPush(sourceSquare, doublePushSquare, piece);
+                    }
+                }
+            }
+
+            var attacks = Attacks.PawnAttacks[(int)position.Side][sourceSquare];
+
+            // En passant
+            if (position.EnPassant != BoardSquare.noSquare && attacks.GetBit(position.EnPassant))
+            // We assume that position.OccupancyBitBoards[oppositeOccupancy].GetBit(targetSquare + singlePush) == true
+            {
+                yield return MoveExtensions.EncodeEnPassant(sourceSquare, (int)position.EnPassant, piece, capturedPiece: (int)Piece.p - offset);
+            }
+
+            // Captures
+            var attackedSquares = attacks & position.OccupancyBitBoards[oppositeSide];
+            while (attackedSquares != default)
+            {
+                targetSquare = attackedSquares.GetLS1BIndex();
+                attackedSquares.ResetLS1B();
+                var capturedPiece = FindCapturedPiece(position, offset, targetSquare);
+
+                var targetRank = (targetSquare >> 3) + 1;
+                if (targetRank == 1 || targetRank == 8)  // Capture with promotion
+                {
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, targetSquare, piece, promotedPiece: (int)Piece.Q + offset, capturedPiece: capturedPiece);
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, targetSquare, piece, promotedPiece: (int)Piece.R + offset, capturedPiece: capturedPiece);
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, targetSquare, piece, promotedPiece: (int)Piece.N + offset, capturedPiece: capturedPiece);
+                    yield return MoveExtensions.EncodePromotion(sourceSquare, targetSquare, piece, promotedPiece: (int)Piece.B + offset, capturedPiece: capturedPiece);
+                }
+                else
+                {
+                    yield return MoveExtensions.EncodeCapture(sourceSquare, targetSquare, piece, capturedPiece: capturedPiece);
+                }
+            }
+        }
+    }
+
+    [Obsolete("Mostly for reference")]
+    internal static IEnumerable<Move> GenerateKingMoves(Position position, bool capturesOnly = false)
+    {
+        var offset = Utils.PieceOffset(position.Side);
+
+        return GeneratePieceMovesForReference((int)Piece.K + offset, position, offset, capturesOnly);
+    }
+
+    [Obsolete("Mostly for reference")]
+    internal static IEnumerable<Move> GenerateKnightMoves(Position position, bool capturesOnly = false)
+    {
+        var offset = Utils.PieceOffset(position.Side);
+
+        return GeneratePieceMovesForReference((int)Piece.N + offset, position, offset, capturesOnly);
+    }
+
+    [Obsolete("Mostly for reference")]
+    internal static IEnumerable<Move> GenerateBishopMoves(Position position, bool capturesOnly = false)
+    {
+        var offset = Utils.PieceOffset(position.Side);
+
+        return GeneratePieceMovesForReference((int)Piece.B + offset, position, offset, capturesOnly);
+    }
+
+    [Obsolete("Mostly for reference")]
+    internal static IEnumerable<Move> GenerateRookMoves(Position position, bool capturesOnly = false)
+    {
+        var offset = Utils.PieceOffset(position.Side);
+
+        return GeneratePieceMovesForReference((int)Piece.R + offset, position, offset, capturesOnly);
+    }
+
+    [Obsolete("Mostly for reference")]
+    internal static IEnumerable<Move> GenerateQueenMoves(Position position, bool capturesOnly = false)
+    {
+        var offset = Utils.PieceOffset(position.Side);
+
+        return GeneratePieceMovesForReference((int)Piece.Q + offset, position, offset, capturesOnly);
+    }
+
+    /// <summary>
+    /// Same as <see cref="GeneratePieceMoves(int, Position, bool)"/> but returning them
+    /// </summary>
+    /// <param name="piece"><see cref="Piece"/></param>
+    /// <param name="position"></param>
+    /// <param name="capturesOnly"></param>
+    /// <returns></returns>
+    [Obsolete("Mostly for reference")]
+    private static IEnumerable<Move> GeneratePieceMovesForReference(int piece, Position position, int offset, bool capturesOnly = false)
+    {
+        var bitboard = position.PieceBitBoards[piece];
+        int sourceSquare, targetSquare;
+
+        while (bitboard != default)
+        {
+            sourceSquare = bitboard.GetLS1BIndex();
+            bitboard.ResetLS1B();
+
+            var attacks = _pieceAttacks[piece](sourceSquare, position.OccupancyBitBoards[(int)Side.Both])
+                & ~position.OccupancyBitBoards[(int)position.Side];
+
+            while (attacks != default)
+            {
+                targetSquare = attacks.GetLS1BIndex();
+                attacks.ResetLS1B();
+
+                if (position.OccupancyBitBoards[(int)Side.Both].GetBit(targetSquare))
+                {
+                    var capturedPiece = FindCapturedPiece(position, offset, targetSquare);
+                    yield return MoveExtensions.EncodeCapture(sourceSquare, targetSquare, piece, capturedPiece: capturedPiece);
+                }
+                else if (!capturesOnly)
+                {
+                    yield return MoveExtensions.Encode(sourceSquare, targetSquare, piece);
+                }
+            }
+        }
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private static int FindCapturedPiece(Position position, int offset, int targetSquare)
+    {
+        var start = (int)Piece.p - offset;
+        for (int pieceIndex = start; pieceIndex < start + 5; ++pieceIndex)
+        {
+            if (position.PieceBitBoards[pieceIndex].GetBit(targetSquare))
+            {
+                return pieceIndex;
+            }
+        }
+
+        throw new AssertException("No captured piece found");
+    }
+}
